fix: move config
