+++ packages/vue-cli-plugin-mpx-web/config/index.js
